scanner/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3
netifaces==0.11.0
python-dotenv==1.0.0
python-nmap==0.7.1
python-frontmatter==1.0.0
